⚡use sub millisecond measurements (#5)
Should increase accuracy a little more when extrapolating out to
schedule phantom beats/ticks in the future.

+++ Apps/AudioAppTemplate/Source/MainComponent.cpp
@@ -11,7 +11,7 @@ namespace AudioApp
         addAndMakeVisible(audioSource);
         addAndMakeVisible(oscComponent);
 
-        tempoAnalyser.onBeat = [this](long long period) {
+        tempoAnalyser.onBeat = [this](double period) {
             oscComponent.sendBeatMessage();
             tempoSynthesizer.beat(period);
         };
